Refactor: Centralize config merge logic in objectUtils
Extracted the 3-way configuration merging logic from `configManagerFactory.js` into a new, reusable function named `mergeWithFileChanges` and moved it to `objectUtils.js`.

This change centralizes the complex merging logic in a single utility file, improving code organization and maintainability. The `configManagerFactory.js` is now cleaner and more focused on its primary role of managing the configuration lifecycle.
